requirements.txt
# Django Framework
Django==5.2.3

# Database
psycopg2-binary==2.9.7  # PostgreSQL (comentar se não usar)

# API Development
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Environment variables
python-decouple==3.8

# Image processing
Pillow==11.2.1

# Development tools
django-debug-toolbar==4.2.0

# Documentation
drf-spectacular==0.26.5

# Timezone support
pytz==2023.3
